+++ API/src/main/java/fr/maxlego08/menu/ComponentsManager.java
@@ -6,9 +6,18 @@ import org.jetbrains.annotations.NotNull;
 import java.util.Optional;
 
 public interface ComponentsManager {
+    /**
+     * Register a new ItemComponentLoader.
+     * @param loader The loader to register
+     * @throws IllegalArgumentException if a loader with the same name is already registered
+     **/
+    void registerComponent(@NotNull ItemComponentLoader loader) throws IllegalArgumentException;
 
-    void registerComponent(@NotNull ItemComponentLoader loader);
-
+    /**
+     * Get an ItemComponentLoader by its name.
+     * @param name The name of the loader
+     * @return An Optional containing the loader if found, or empty if not found
+     **/
     @NotNull
-    Optional<ItemComponentLoader> getLoader(@NotNull String key) throws IllegalArgumentException;
+    Optional<ItemComponentLoader> getLoader(@NotNull String name);
 }