+++ changelog.md
@@ -42,6 +42,9 @@
 
 # Unreleased
 
+- New component item system for 1.20.5+
+- Animated title (Required `PacketEvent`)
+- Added automatic support for newer Minecraft versions when no major API changes are present
 - Fix Avast flag issue on VirusTotal
 - Introduced the actions_patterns system, which lets you define default actions applied to all buttons unless they already specify those action types.
 - MiniMessage support in toasts.